COMMERCE-8276 dispatch-test - reorganize - move to different package
